Former-commit-id: f44f4d9c91a1d26bdf5360059472dd1eac929edd

+++ plugin-resources/webapp/xnat/java/org/nrg/xdat/om/base/BaseXnatProjectdata.java
@@ -69,18 +69,19 @@ import org.nrg.xft.exception.DBPoolException;
 import org.nrg.xft.exception.ElementNotFoundException;
 import org.nrg.xft.exception.FieldNotFoundException;
 import org.nrg.xft.exception.InvalidItemException;
+import org.nrg.xft.exception.InvalidPermissionException;
 import org.nrg.xft.exception.InvalidValueException;
 import org.nrg.xft.exception.XFTInitException;
 import org.nrg.xft.schema.Wrappers.GenericWrapper.GenericWrapperElement;
 import org.nrg.xft.search.ItemSearch;
 import org.nrg.xft.security.UserI;
 import org.nrg.xft.utils.FileUtils;
+import org.nrg.xft.utils.SaveItemHelper;
 import org.nrg.xft.utils.StringUtils;
 import org.nrg.xnat.exceptions.InvalidArchiveStructure;
 import org.nrg.xnat.helpers.prearchive.PrearcUtils;
 import org.nrg.xnat.turbine.utils.ArcSpecManager;
 import org.nrg.xnat.turbine.utils.ArchivableItem;
-import org.restlet.data.Status;
 
 /**
  * @author XDAT
@@ -877,18 +878,26 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 
     public String addGroupMember(String group_id, XDATUser newUser,XDATUser currentUser) throws Exception{
 		final String confirmquery = "SELECT * FROM xdat_user_groupid WHERE groupid='" + group_id + "' AND groups_groupid_xdat_user_xdat_user_id=" + newUser.getXdatUserId() + ";";
-    	XFTTable t=XFTTable.Execute(confirmquery,currentUser.getDBName(), currentUser.getUsername());
+    	if(!currentUser.canDelete(this)){
+    		throw new InvalidPermissionException("User cannot modify project " + this.getId());
+    	}
+		
+		XFTTable t=XFTTable.Execute(confirmquery,currentUser.getDBName(), currentUser.getUsername());
     	if(t.size()==0){
             final XdatUserGroupid map = new XdatUserGroupid((UserI)currentUser);
             map.setProperty(map.getXSIType() +".groups_groupid_xdat_user_xdat_user_id", newUser.getXdatUserId());
             map.setGroupid(group_id);
-            map.save(currentUser, false, false);
+            SaveItemHelper.authorizedSave(map,currentUser, false, false);
     	}
         return group_id;
     }
 
     public void removeGroupMember(String group_id, XDATUser newUser,XDATUser currentUser) throws Exception{
-		final String confirmquery = "SELECT * FROM xdat_user_groupid WHERE groupid='" + group_id + "' AND groups_groupid_xdat_user_xdat_user_id=" + newUser.getXdatUserId() + ";";
+    	if(!currentUser.canDelete(this)){
+    		throw new InvalidPermissionException("User cannot modify project " + this.getId());
+    	}
+    	
+    	final String confirmquery = "SELECT * FROM xdat_user_groupid WHERE groupid='" + group_id + "' AND groups_groupid_xdat_user_xdat_user_id=" + newUser.getXdatUserId() + ";";
     	XFTTable t=XFTTable.Execute(confirmquery,currentUser.getDBName(), currentUser.getUsername());
     	if(t.size()>0){
     		final String query = "DELETE FROM xdat_user_groupid WHERE groupid='" + group_id + "' AND groups_groupid_xdat_user_xdat_user_id=" + newUser.getXdatUserId() + ";";
@@ -917,7 +926,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
             logger.error("",e2);
         }
         if (XFT.VERBOSE)System.out.println("Group init() BEGIN: " + (Calendar.getInstance().getTimeInMillis()-startTime) + "ms");
-
+        
         for (final ElementSecurity es:(ArrayList<ElementSecurity>)ess.clone())
         {
             try {
@@ -943,7 +952,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                 group.setId(getId() + "_" + OWNER_GROUP);
                 group.setDisplayname("Owners");
                 group.setTag(getId());
-                group.save(this.getUser(), true, true);
+                SaveItemHelper.authorizedSave(group,this.getUser(), true, true);
             } catch (Exception e) {
                 logger.error("",e);
             }
@@ -996,7 +1005,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                 group.setId(getId() + "_" + MEMBER_GROUP);
                 group.setDisplayname("Members");
                 group.setTag(getId());
-                group.save(this.getUser(), true, true);
+                SaveItemHelper.authorizedSave(group,this.getUser(), true, true);
             } catch (Exception e) {
                 logger.error("",e);
             }
@@ -1043,7 +1052,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                 group.setId(getId() + "_" + COLLABORATOR_GROUP);
                 group.setDisplayname("Collaborators");
                 group.setTag(getId());
-                group.save(this.getUser(), true, true);
+                SaveItemHelper.authorizedSave(group,this.getUser(), true, true);
             } catch (Exception e) {
                 logger.error("",e);
             }
@@ -1423,7 +1432,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                         XFTItem i = XFTItem.NewItem("xdat:user",this.getUser());
                         i.setProperty("login","guest");
                         i.setProperty("enabled", 0);
-                        i.save(this.getUser(), true, false);
+                        SaveItemHelper.authorizedSave(i, this.getUser(), true, false);
 
                         guest = new XDATUser("guest");
                     }
@@ -1499,7 +1508,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 
         UserGroup ownerG=UserGroupManager.GetGroup(group.getId());
         if(ownerG==null){
-        	group.save(this.getUser(), true, true);
+        	SaveItemHelper.authorizedSave(group,this.getUser(), true, true);
 
             group.setPermissions("xnat:projectData", "xnat:projectData/ID", getId(), Boolean.TRUE, Boolean.TRUE, Boolean.TRUE, Boolean.TRUE, Boolean.TRUE, Boolean.TRUE, (XDATUser)user,false);
 
@@ -1666,6 +1675,10 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
     }
 
     public void initArcProject(ArcProject arcP, XDATUser user) throws Exception{
+    	if(!user.canDelete(this)){
+    		throw new InvalidPermissionException("User cannot modify project " + this.getId());
+    	}
+    	
     	if(arcP==null){
     		 XFTItem item = XFTItem.NewItem("arc:project", user);
     	     arcP = new ArcProject(item);
@@ -1679,7 +1692,8 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
         arcP.setProperty("arc:project/paths/cachePath", ArcSpecManager.GetInstance().getGlobalCachePath() + getId()+"/");
         arcP.setProperty("arc:project/paths/buildPath", ArcSpecManager.GetInstance().getGlobalBuildPath() + getId()+"/");
 
-        arcP.save(user, true, false);
+        
+        SaveItemHelper.authorizedSave(arcP,user, true, false);
         ArcSpecManager.Reset();
     }
 
@@ -1889,6 +1903,11 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
     public void delete(boolean removeFiles, XDATUser user)throws SQLException,Exception{
     	boolean preventProjectDelete=false;
         boolean preventProjectDeleteByP=false;
+        
+        if(!user.canDelete(this)){
+        	throw new InvalidPermissionException("User cannot delete project:" + getId());
+        }
+        
     	for (XnatSubjectdata subject : getParticipants_participant()){
             if (subject!=null){
             	boolean preventSubjectDelete=false;
@@ -1918,7 +1937,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                                 }
                             }
 
-                        	DBAction.DeleteItem(expt.getItem().getCurrentDBVersion(), user);
+                            SaveItemHelper.authorizedDelete(expt.getItem().getCurrentDBVersion(), user);
                         }else{
                         	preventSubjectDeleteByP=true;
                         }
@@ -1926,7 +1945,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                     	preventSubjectDelete=true;
                     	for(XnatExperimentdataShareI pp : expt.getSharing_share()){
                     		if(pp.getProject().equals(getId())){
-                    			DBAction.DeleteItem(((XnatExperimentdataShare)pp).getItem(),user);
+                    			SaveItemHelper.authorizedDelete(((XnatExperimentdataShare)pp).getItem(),user);
                     		}
                     	}
                     }
@@ -1937,7 +1956,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
             	if(!subject.getProject().equals(getId())){
             		for(XnatProjectparticipantI pp : subject.getSharing_share()){
                 		if(pp.getProject().equals(getId())){
-                			DBAction.DeleteItem(((XnatExperimentdataShare)pp).getItem(),user);
+                			SaveItemHelper.authorizedDelete(((XnatExperimentdataShare)pp).getItem(),user);
                 		}
                 	}
             	}else{
@@ -1958,7 +1977,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
                                     }
                                 }
                             }
-                    		DBAction.DeleteItem(subject.getItem().getCurrentDBVersion(), user);
+                            SaveItemHelper.authorizedDelete(subject.getItem().getCurrentDBVersion(), user);
                 		}else{
                 			preventProjectDeleteByP=true;
                 		}
@@ -1975,7 +1994,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 			final File prearc=new File(this.getPrearchivePath());
 			
 			PrearcUtils.deleteProject(this.getId());
-	        DBAction.DeleteItem(getItem().getCurrentDBVersion(), user);
+			SaveItemHelper.authorizedDelete(getItem().getCurrentDBVersion(), user);
 
 	        //DELETE field mappings
 	        ItemSearch is = ItemSearch.GetItemSearch("xdat:field_mapping", user);
@@ -1984,7 +2003,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 	        while (items.hasNext())
 	        {
 	            XFTItem item = (XFTItem)items.next();
-	            DBAction.DeleteItem(item, user);
+	            SaveItemHelper.authorizedDelete(item, user);
 	        }
 
 	        //DELETE user.groupId
@@ -1995,7 +2014,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 	        while(groups.hasNext()){
 	            XdatUserGroupid g = (XdatUserGroupid)groups.next();
 	            try {
-	                DBAction.DeleteItem(g.getItem(), user);
+	            	SaveItemHelper.authorizedDelete(g.getItem(), user);
 	            } catch (Throwable e) {
 	                logger.error("",e);
 	            }
@@ -2009,7 +2028,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 	        while(groups.hasNext()){
 	            XdatUsergroup g = (XdatUsergroup)groups.next();
 	            try {
-	                DBAction.DeleteItem(g.getItem(), user);
+	            	SaveItemHelper.authorizedDelete(g.getItem(), user);
 	            } catch (Throwable e) {
 	                logger.error("",e);
 	            }
@@ -2021,7 +2040,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 	        {
 	            ItemI bundle = (ItemI)bundles.next();
 	            try {
-	                DBAction.DeleteItem(bundle.getItem(), user);
+	            	SaveItemHelper.authorizedDelete(bundle.getItem(), user);
 	            } catch (Throwable e) {
 	                logger.error("",e);
 	            }
@@ -2029,7 +2048,7 @@ public class BaseXnatProjectdata extends AutoXnatProjectdata  implements Archiva
 
 	        ArcProject p =getArcSpecification();
 	        try {
-	            if (p!=null)DBAction.DeleteItem(p.getItem(), user);
+	            if (p!=null)SaveItemHelper.authorizedDelete(p.getItem(), user);
 	        } catch (Throwable e) {
 	            logger.error("",e);
 	        }
